+++ src/utils/constants/configDefaults.ts
@@ -1,10 +1,8 @@
 const configDefaults = {
 	ABS_PATH: '',
-	REL_PATH: '',
-	FILE_PREFIX: '',
-	LANGUAGE: 'en-us',
-	FILE_SUFFIX: '',
-	FILE_SEPARATOR: '.',
+	ENABLED: false,
+	FILE: '',
+	LANGUAGES: ['en-us'],
 	SHOW_FLAG: false,
 	WATCH_FILES: true
 };