+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
 
+## Version 0.0.5-alpha
+* Easier configuration
+* Supports multiple term languages
+* On/Off extension toggle
+
 ## Version 0.0.1-alpha
 * First release